--- a/mopro-msm/src/msm/metal_msm/shader/curve/jacobian_madd_2007_bl.metal
+++ b/mopro-msm/src/msm/metal_msm/shader/curve/jacobian_madd_2007_bl.metal
@@ -16,17 +16,17 @@ kernel void run(
     device BigInt* result_zr [[ buffer(7) ]],
     uint gid [[ thread_position_in_grid ]]
 ) {
-    FieldElement x1 = FieldElement{ *a_xr };
-    FieldElement y1 = FieldElement{ *a_yr };
-    FieldElement z1 = FieldElement{ *a_zr };
-    FieldElement x2 = FieldElement{ *b_xr };
-    FieldElement y2 = FieldElement{ *b_yr };
+    BigInt x1 = *a_xr;
+    BigInt y1 = *a_yr;
+    BigInt z1 = *a_zr;
+    BigInt x2 = *b_xr;
+    BigInt y2 = *b_yr;
 
-    Jacobian a = Jacobian{ .x = x1, .y = y1, .z = z1 };
-    Affine b = Affine{ .x = x2, .y = y2 };
+    Jacobian a; a.x = x1; a.y = y1; a.z = z1;
+    Affine b; b.x = x2; b.y = y2;
 
-    Jacobian res = a + b;
-    *result_xr = res.x.value;
-    *result_yr = res.y.value;
-    *result_zr = res.z.value;
+    Jacobian res = jacobian_madd_2007_bl(a, b);
+    *result_xr = res.x;
+    *result_yr = res.y;
+    *result_zr = res.z;
 }