--- a/mopro-msm/src/msm/metal_msm/shader/curve/utils.metal
+++ b/mopro-msm/src/msm/metal_msm/shader/curve/utils.metal
@@ -7,14 +7,14 @@ using namespace metal;
 #include "../misc/get_constant.metal"
 
 bool is_jacobian_zero(Jacobian a) {
-    return is_bigint_zero(a.z.value);
+    return is_bigint_zero(a.z);
 }
 
 bool jacobian_eq(Jacobian lhs, Jacobian rhs) {
     for (uint i = 0; i < NUM_LIMBS; i++) {
-        if (lhs.x.value.limbs[i] != rhs.x.value.limbs[i]) return false;
-        else if (lhs.y.value.limbs[i] != rhs.y.value.limbs[i]) return false;
-        else if (lhs.z.value.limbs[i] != rhs.z.value.limbs[i]) return false;
+        if (lhs.x.limbs[i] != rhs.x.limbs[i]) return false;
+        else if (lhs.y.limbs[i] != rhs.y.limbs[i]) return false;
+        else if (lhs.z.limbs[i] != rhs.z.limbs[i]) return false;
     }
     return true;
 }