--- a/mopro-msm/src/msm/metal_msm/shader/curve/utils.metal
+++ b/mopro-msm/src/msm/metal_msm/shader/curve/utils.metal
@@ -6,11 +6,11 @@ using namespace metal;
 #include "../bigint/bigint.metal"
 #include "../misc/get_constant.metal"
 
-bool is_jacobian_zero(Jacobian a) {
+inline bool is_jacobian_zero(Jacobian a) {
     return is_bigint_zero(a.z);
 }
 
-bool jacobian_eq(Jacobian lhs, Jacobian rhs) {
+inline bool jacobian_eq(Jacobian lhs, Jacobian rhs) {
     for (uint i = 0; i < NUM_LIMBS; i++) {
         if (lhs.x.limbs[i] != rhs.x.limbs[i]) return false;
         else if (lhs.y.limbs[i] != rhs.y.limbs[i]) return false;