- Closes issue: #60

## Summary
This PR reverts the use of `FieldElement` (introduced in commit f6daf7a7458b12b8ac437530807ab6e3b51f3ebb) back to `BigInt`. This change significantly reduces load/store instructions and register pressure, especially in performance-critical tight loops.

Additionally, `inline` attributes were added to frequently invoked functions, further decreasing function call overhead and improving overall performance.

## Montgomery Multiplication Benchmarks
Benchmarks were conducted on a system with:
- Apple M3 chip
- 24 GB unified memory
- macOS 15
- Metal 32023.404

### Benchmark Command
```
cargo test --package mopro-msm --lib -- msm::metal_msm::tests --ignored --nocapture
```

### Benchmark Results

#### Before Optimization
```
=== benchmarking mont_mul_modified ===
- 13-bit limbs: 251ms
- 15-bit limbs: 260ms

=== benchmarking mont_mul_optimised ===
- 13-bit limbs: 239ms

=== benchmarking mont_mul_cios ===
- 13-bit limbs: 827ms
- 15-bit limbs: 625ms
- 16-bit limbs: 541ms
```

#### After Optimization
```
=== benchmarking mont_mul_modified ===
- 13-bit limbs: 189ms (~25% improvement)
- 15-bit limbs: 198ms (~24% improvement)

=== benchmarking mont_mul_optimised ===
- 13-bit limbs: 195ms (~18% improvement)

=== benchmarking mont_mul_cios ===
- 13-bit limbs: 430ms (~48% improvement)
- 15-bit limbs: 334ms (~46% improvement)
- 16-bit limbs: 185ms (~66% improvement)
```

This PR significantly improves performance across various limb sizes for Montgomery multiplication.